backend/requirements.txt
# FastAPI & Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
SQLAlchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

# ML Inference
Pillow>=10.0.0
numpy>=1.24.0
onnxruntime>=1.16.0

# Monitoring & Scheduling
apscheduler>=3.10.0
scikit-learn>=1.3.0  # for metric_service F1/accuracy computation

# Testing & Linting
pytest>=7.4.0
httpx>=0.24.0
pytest-cov>=4.1.0
